+++ R/data.R
@@ -118,6 +118,15 @@
 #' @seealso \code{\link{tlBooks}} \code{\link{tlEvents}}
 "tlFootnotes"
 
+#' Star Trek logos metadata.
+#'
+#' A data frame with 236 rows and 3 columns containing Star Trek logo metadata: category, description and URL. Logo artwork credited to Kris Trigwell.
+#' The logo images are served by st-minutiae.com for personal and fair use.
+#'
+#' @format A data frame
+#' @seealso \code{\link{st_logo}}
+"stLogos"
+
 #' Available datasets
 #'
 #' List the available datasets in the rtrek package.
@@ -131,7 +140,7 @@ st_datasets <- function(){
   dplyr::tibble(
     dataset = c("stGeo", "stSpecies", "stTiles",
                 "stBooks", "stBooksWiki", "stSeries",
-                "stapiEntities",
+                "stapiEntities", "stLogos",
                 "tlBooks", "tlEvents", "tlFootnotes"),
     description = c("Map tile set locations of interest.",
                     "Basic intelligent species data.",
@@ -140,6 +149,7 @@ st_datasets <- function(){
                     "Star Trek novel metadata from Wikipedia.",
                     "Names and acronyms of Star Trek series",
                     "Star Trek API (STAPI) categories",
+                    "Metadata for various Star Trek logos",
                     "Novel-based timeline dataset",
                     "Event-based timeline dataset",
                     "Timeline dataset footnotes"))