+++ R/ma_support.R
@@ -40,7 +40,8 @@ ma_select <- memoise::memoise(ma_select)
 # Recursively collate all subcategories or articles for a given category's page(s)
 ma_category_pages <- function(id, url, nodes, d0 = NULL){
   x <- xml2::read_html(ma_base_add(url))
-  x1 <- rvest::html_nodes(x, nodes[1]) %>% rvest::html_nodes("a")
+  x1 <- rvest::html_nodes(x, nodes[1]) %>% rvest::html_nodes("ul a")
+  x1 <- x1[grepl("class=\"category-page__member-link\"", x1)]
   txt <- ma_text(x1) %>% ma_strip_prefix()
   url <- ma_href(x1)
   d <- dplyr::tibble(txt, url) %>% stats::setNames(c(id, "url"))