+++ R/zzz.R
@@ -1,5 +1,5 @@
 .onLoad <- function(libname, pkgname){
-  options(rtrek_antiddos = 1)
+  options(rtrek_antidos = 1)
 }
 
 rtrek_api_time <- new.env()